projectMQTTsub/requirements.txt
paho-mqtt
httpx